updated polynomial.py to include the changes made to multi_cheb.py and multi_power.py
