+++ groebner/polynomial.py
@@ -203,8 +203,6 @@ class Polynomial(object):
 
 #### MULTI_CHEB ###############################################################
 
-
-
 class MultiCheb(Polynomial):
     """
     _____ params _______
@@ -221,10 +219,10 @@ class MultiCheb(Polynomial):
         input- Current: list, current location in ordering
         output- the next step in ordering
     """
-    
+
     def printTime():
         print(times)
-    
+
     def clearTime():
         times["mon_mult_cheb"] = 0
 
@@ -344,26 +342,45 @@ class MultiCheb(Polynomial):
         return sol
 
     def mon_mult(self, idx, returnType = 'Poly'):
+        """
+        Multiplies a Chebyshev polynomial by a monomial
+        -------
+        Parameters:
+            self: A MultiCheb object
+            idx: The index of the monomial to multiply self by.
+            returnType: if 'Poly' then returns a polynomial object
+        -------
+        Returns:
+            MultiCheb object or a matrix if returnType is not 'Poly'
+        -------
+        """
         start = time.time()
-        
         initial_matrix = self.coeff
         for i in range(len(idx)):
             idx_zeros = np.zeros(len(idx),dtype = int)
             idx_zeros[i] = idx[i]
-            initial_matrix = MultiCheb.mon_mult1(initial_matrix, idx_zeros)
+            initial_matrix = MultiCheb.mon_mult1(initial_matrix, idx_zeros, i)
         if returnType == 'Poly':
-            poly = MultiCheb(initial_matrix, lead_term = self.lead_term + np.array(idx), clean_zeros = False)
             end = time.time()
-            times["mon_mult_cheb"] += (end-start)
-            return poly
+            times["mon_mult_cheb"] += (end - start)
+            return MultiCheb(initial_matrix, lead_term = self.lead_term + np.array(idx), clean_zeros = False)
         elif returnType == 'Matrix':
             end = time.time()
-            times["mon_mult_cheb"] += (end-start)
+            times["mon_mult_cheb"] += (end - start)
             return initial_matrix
 
-    def mon_mult1(initial_matrix, idx):
+    def mon_mult1(initial_matrix, idx, dim_mult):
         """
-        Takes a polynomial and the index of a monomial and returns the result of the multiplication.
+        Executes monomial multiplication in one dimension
+        -------
+        Parameters:
+            initial_matrix: matrix of coefficients that represents a Chebyshev polynomial
+            idx: the index of a monomial of one variable to multiply the Chebyshev polynomial by
+            dim_mult: the location of the non-zero value in idx.
+        -------
+        Returns:
+            matrix of coeff that is the result of the one dimensial monomial multiplication.
+        -------
         """
         pad_values = list()
         for i in idx: #iterates through monomial and creates a tuple of pad values for each dimension
@@ -374,12 +391,13 @@ class MultiCheb(Polynomial):
 
         largest_idx = [i-1 for i in initial_matrix.shape]
         new_shape = [max(i,j) for i,j in itertools.zip_longest(largest_idx, idx, fillvalue = 0)] #finds the largest length in each dimmension
-        add_a = [i-j for i,j in itertools.zip_longest(new_shape, largest_idx, fillvalue = 0)]
-        add_a_list = np.zeros((len(new_shape),2))
-        #changes the second column to the values of add_a and add_b.
-        add_a_list[:,1] = add_a
-        #uses add_a_list and add_b_list to pad each polynomial appropriately.
-        initial_matrix = np.pad(initial_matrix,add_a_list.astype(int),'constant')
+        if initial_matrix.shape[dim_mult] <= idx[dim_mult]:
+            add_a = [i-j for i,j in itertools.zip_longest(new_shape, largest_idx, fillvalue = 0)]
+            add_a_list = np.zeros((len(new_shape),2))
+            #changes the second column to the values of add_a and add_b.
+            add_a_list[:,1] = add_a
+            #uses add_a_list and add_b_list to pad each polynomial appropriately.
+            initial_matrix = np.pad(initial_matrix,add_a_list.astype(int),'constant')
 
         number_of_dim = initial_matrix.ndim
         shape_of_self = initial_matrix.shape
@@ -422,7 +440,6 @@ class MultiCheb(Polynomial):
 
 #### MULTI_POWER ##############################################################
 
-
 class MultiPower(Polynomial):
     """
     _____ params _______
@@ -441,13 +458,13 @@ class MultiPower(Polynomial):
     """
     def printTime():
         print(times)
-    
+
     def clearTime():
         times["mon_mult_power"] = 0
 
     def __init__(self, coeff, order='degrevlex', lead_term=None, clean_zeros = True):
         super(MultiPower, self).__init__(coeff, order, lead_term, clean_zeros)
-        
+
     def __add__(self,other):
         '''
         Here we add an addition class.
@@ -506,10 +523,6 @@ class MultiPower(Polynomial):
         for i in M:
             list1 = (i,0)
             tuple1.append(list1)
-        poly = MultiPower(np.pad(self.coeff, tuple1, 'constant', constant_values = 0), 
-                          clean_zeros = False, lead_term = self.lead_term + M)
-        end = time.time()
-        times["mon_mult_power"] += (end-start)
         if returnType == 'Poly':
             poly = MultiPower(np.pad(self.coeff, tuple1, 'constant', constant_values = 0), 
                           clean_zeros = False, lead_term = self.lead_term + M)
@@ -538,7 +551,6 @@ class MultiPower(Polynomial):
             return 0
         else:
             return poly_value
-
 
 ###############################################################################
